@@ -1,220 +1,270 @@
-﻿using System;
+﻿using System.Diagnostics.Contracts;
+using System.Numerics;
 
-namespace Rationals
+namespace Rationals;
+
+/// <summary>
+///     Rational number.
+/// </summary>
+public struct Rational : IComparable<Rational>, IEquatable<Rational>
 {
-    public struct Rational
+    /// <summary>
+    ///     Ctor.
+    /// </summary>
+    /// <param name="numerator">Numerator</param>
+    /// <param name="denominator">Denominator</param>
+    /// <exception cref="DivideByZeroException">Throws when denominator is zero</exception>
+    public Rational(BigInteger numerator, BigInteger denominator)
     {
-        /// <summary>
-        /// Constructs rational Numeratorber out of a whole Numeratorber.
-        /// </summary>
-        public Rational(int whole)
-            : this(whole, 1)
+        if (denominator.IsZero)
+            throw new DivideByZeroException("Denominator cannot be zero.");
+
+        Numerator = numerator;
+        Denominator = denominator;
+    }
+    
+    /// <summary>
+    ///     Ctor.
+    /// </summary>
+    /// <param name="numerator">Numerator</param>
+    public Rational(BigInteger numerator) : this(numerator, BigInteger.One)
+    {
+    }
+
+    /// <summary>
+    ///     Simplifies the numerator and denominator of a rational number.
+    /// </summary>
+    /// <returns>Reference to simplified version of a rational number.</returns>
+    public Rational Simplify()
+    {
+        var gcd = BigInteger.GreatestCommonDivisor(Numerator, Denominator);
+        return new Rational(Numerator / gcd, Denominator / gcd);
+    }
+    
+    /// <summary>
+    ///     Numerator part of the rational numerator.
+    /// </summary>
+    public BigInteger Numerator { get; }
+
+    /// <summary>
+    ///     Denominator part of the rational number.
+    /// </summary>
+    public BigInteger Denominator { get; }
+
+    /// <summary>
+    ///     True if the number is equal to zero.
+    /// </summary>
+    public bool IsZero => !IsNaN && Numerator.IsZero;
+
+    /// <summary>
+    ///     True if the number is equal to one.
+    /// </summary>
+    public bool IsOne => !IsNaN && Numerator == Denominator;
+
+    /// <summary>
+    ///     True if the value is not a number.
+    /// </summary>
+    public bool IsNaN => Denominator.IsZero;
+
+    /// <summary>
+    ///     Gets a number that indicates the sign (negative, positive, or zero) of the rational number.
+    /// </summary>
+    public int Sign => IsNaN ? 0 : Numerator.Sign * Denominator.Sign;
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="a"></param>
+    /// <returns></returns>
+    public static Rational operator +(Rational a)
+    {
+        return a;
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="a"></param>
+    /// <returns></returns>
+    public static Rational operator -(Rational a)
+    {
+        return new Rational(-a.Numerator, a.Denominator);
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="a"></param>
+    /// <param name="b"></param>
+    /// <returns></returns>
+    public static Rational operator +(Rational a, Rational b)
+    {
+        return new Rational(a.Numerator * b.Denominator + b.Numerator * a.Denominator,
+            a.Denominator * b.Denominator);
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="a"></param>
+    /// <param name="b"></param>
+    /// <returns></returns>
+    public static Rational operator -(Rational a, Rational b)
+    {
+        return a + -b;
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="a"></param>
+    /// <param name="b"></param>
+    /// <returns></returns>
+    public static Rational operator *(Rational a, Rational b)
+    {
+        return new Rational(a.Numerator * b.Numerator, a.Denominator * b.Denominator);
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="a"></param>
+    /// <param name="b"></param>
+    /// <returns></returns>
+    /// <exception cref="DivideByZeroException"></exception>
+    public static Rational operator /(Rational a, Rational b)
+    {
+        if (b.Numerator == 0)
+            throw new DivideByZeroException("Denominator cannot be zero.");
+        return new Rational(a.Numerator * b.Denominator, a.Denominator * b.Numerator);
+    }
+
+    /// <summary>
+    ///     Overload of the ++ operator.
+    /// </summary>
+    /// <param name="a">Rational number.</param>
+    /// <returns>Rational with +1.</returns>
+    public static Rational operator ++(Rational a)
+    {
+        return a + new Rational(1);
+    }
+
+    /// <summary>
+    ///     Overload of the -- operator.
+    /// </summary>
+    /// <param name="a">Rational number.</param>
+    /// <returns>Rational with -1.</returns>
+    public static Rational operator --(Rational a)
+    {
+        return a - new Rational(1);
+    }
+
+    /// <summary>
+    ///     Overload of the == operator.
+    /// </summary>
+    /// <param name="left"></param>
+    /// <param name="right"></param>
+    /// <returns></returns>
+    public static bool operator ==(Rational left, Rational right)
+    {
+        return !left.IsNaN && !right.IsNaN && left.Equals(right);
+    }
+
+    /// <summary>
+    ///     Overload of the != operator.
+    /// </summary>
+    /// <param name="left"></param>
+    /// <param name="right"></param>
+    /// <returns></returns>
+    public static bool operator !=(Rational left, Rational right)
+    {
+        return !(left == right);
+    }
+
+    /// <summary>
+    ///     Overload of the &lt; operator.
+    /// </summary>
+    /// <param name="left"></param>
+    /// <param name="right"></param>
+    /// <returns></returns>
+    public static bool operator <(Rational left, Rational right)
+    {
+        return !left.IsNaN && !right.IsNaN && left.CompareTo(right) < 0;
+    }
+
+    /// <summary>
+    ///     Overload of the &gt; operator.
+    /// </summary>
+    /// <param name="left"></param>
+    /// <param name="right"></param>
+    /// <returns></returns>
+    public static bool operator >(Rational left, Rational right)
+    {
+        return !left.IsNaN && !right.IsNaN && left.CompareTo(right) > 0;
+    }
+
+    /// <summary>
+    ///     Overload of the &lt;= operator.
+    /// </summary>
+    /// <param name="left"></param>
+    /// <param name="right"></param>
+    /// <returns></returns>
+    public static bool operator <=(Rational left, Rational right)
+    {
+        return !left.IsNaN && !right.IsNaN && left.CompareTo(right) <= 0;
+    }
+
+    /// <summary>
+    ///     Overload of the &gt;= operator.
+    /// </summary>
+    /// <param name="left"></param>
+    /// <param name="right"></param>
+    /// <returns></returns>
+    public static bool operator >=(Rational left, Rational right)
+    {
+        return !left.IsNaN && !right.IsNaN && left.CompareTo(right) >= 0;
+    }
+
+    public bool Equals(Rational other)
+    {
+        if (IsNaN)
+            return other.IsNaN;
+
+        return !other.IsNaN && (Numerator * other.Denominator).Equals(other.Numerator * Denominator);
+    }
+
+    public override string ToString()
+    {
+        this = Simplify();
+        return $"{Numerator} / {Denominator}";
+    }
+
+    public int CompareTo(Rational other)
+    {
+        if (IsNaN) return other.IsNaN ? 0 : -1;
+
+        if (other.IsNaN) return 1;
+
+        if (Sign == other.Sign)
         {
+            var adjDenominator = BigInteger.Abs(Denominator) * other.Denominator.Sign;
+            var adjOtherDenominator = BigInteger.Abs(other.Denominator) * Denominator.Sign;
+            return (Numerator * adjOtherDenominator).CompareTo(other.Numerator * adjDenominator);
         }
 
-        /// <summary>
-        /// Constructs rational numerator out of numerator and denominator.
-        /// </summary>
-        public Rational(int numerator, int denominator)
-        {
-            if (denominator == 0)
-                throw new DivideByZeroException("Denominator cannot be zero.");
+        if (Sign > other.Sign)
+            return 1;
 
-            Numerator = numerator;
-            Denominator = denominator;
-        }
+        return -1;
+    }
 
-        /// <summary>
-        /// Numerator part of the rational numerator.
-        /// </summary>
-        public int Numerator { get; set; }
+    public override bool Equals(object? obj)
+    {
+        return obj is Rational other && Equals(other);
+    }
 
-        /// <summary>
-        /// Denominator part of the rational number.
-        /// </summary>
-        public int Denominator { get; set; }
-
-        /// <summary>
-        /// True if the number is equal to zero.
-        /// </summary>
-        public bool IsZero => !IsNaN && Numerator == 0;
-
-        /// <summary>
-        /// True if the number is equal to one.
-        /// </summary>
-        public bool IsOne => !IsNaN && Numerator == Denominator;
-
-        /// <summary>
-        /// True if the value is not a number.
-        /// <see cref="NaN"/>
-        /// </summary>
-        public bool IsNaN => Denominator == 0;
-
-        /// <summary>
-        /// Gets a number that indicates the sign (negative, positive, or zero) of the rational number.
-        /// </summary>
-        public int Sign => IsNaN ? 0 : Math.Sign(Numerator) * Math.Sign(Denominator);
-
-
-        public static Rational operator +(Rational a) => a;
-        public static Rational operator -(Rational a) => new(-a.Numerator, a.Denominator);
-
-        public static Rational operator +(Rational a, Rational b)
-        => new(a.Numerator * b.Denominator + b.Numerator * a.Denominator, a.Denominator * b.Denominator);
-
-        public static Rational operator -(Rational a, Rational b)
-            => a + (-b);
-
-        public static Rational operator *(Rational a, Rational b)
-            => new(a.Numerator * b.Numerator, a.Denominator * b.Denominator);
-
-        public static Rational operator /(Rational a, Rational b)
-        {
-            if (b.Numerator == 0)
-                throw new DivideByZeroException("Denominator cannot be zero.");
-            return new(a.Numerator * b.Denominator, a.Denominator * b.Numerator);
-        }
-
-        /// <summary>
-        /// Checks equalty with other rational number
-        /// </summary>
-        /// <param name="other">Rational number</param>
-        /// <returns>Test of equalty</returns>
-        public bool Equals(Rational other)
-        {
-            if (IsNaN)
-                return other.IsNaN;
-
-            if (other.IsNaN)
-                return false;
-
-            return (Numerator * other.Denominator).Equals(other.Numerator * Denominator);
-        }
-
-        /// <summary>
-        /// Compares rational numbers with each other
-        /// </summary>
-        /// <param name="other">Rational number</param>
-        /// <returns></returns>
-        public int CompareTo(Rational other)
-        {
-            if (IsNaN)
-            {
-                return other.IsNaN ? 0 : -1;
-            }
-
-            if (other.IsNaN)
-            {
-                return 1;
-            }
-
-            if (Sign == other.Sign)
-            {
-
-                var adjDenominator = Math.Abs(Denominator) * Math.Sign(other.Denominator);
-                var adjOtherDenominator = Math.Abs(other.Denominator) * Math.Sign(Denominator);
-                return (Numerator * adjOtherDenominator).CompareTo(other.Numerator * adjDenominator);
-            }
-
-            if (Sign > other.Sign)
-                return 1;
-
-            return -1;
-        }
-
-        /// <summary>
-        /// Overload of the ++ operator.
-        /// </summary>
-        public static Rational operator ++(Rational a) => a + new Rational(1);
-
-        /// <summary>
-        /// Overload of the -- operator.
-        /// </summary>
-        public static Rational operator --(Rational a) => a - new Rational(1);
-
-        /// <summary>
-        /// Overload of the == operator.
-        /// </summary>
-        public static bool operator ==(Rational left, Rational right) => !left.IsNaN && !right.IsNaN && left.Equals(right);
-
-        /// <summary>
-        /// Overload of the != operator.
-        /// </summary>
-        public static bool operator !=(Rational left, Rational right) => !(left == right);
-
-        /// <summary>
-        /// Overload of the &lt; operator.
-        /// </summary>
-        public static bool operator <(Rational left, Rational right) => !left.IsNaN && !right.IsNaN && left.CompareTo(right) < 0;
-
-        /// <summary>
-        /// Overload of the &gt; operator.
-        /// </summary>
-        public static bool operator >(Rational left, Rational right) => !left.IsNaN && !right.IsNaN && left.CompareTo(right) > 0;
-
-        /// <summary>
-        /// Overload of the &lt;= operator.
-        /// </summary>
-        public static bool operator <=(Rational left, Rational right) => !left.IsNaN && !right.IsNaN && left.CompareTo(right) <= 0;
-
-        /// <summary>
-        /// Overload of the &gt;= operator.
-        /// </summary>
-        public static bool operator >=(Rational left, Rational right) => !left.IsNaN && !right.IsNaN && left.CompareTo(right) >= 0;
-
-
-        /// <summary>
-        /// Simplifies rational number
-        /// </summary>
-        private void Simplify()
-        {
-            var gcd = Gcd(Numerator, Denominator);
-            Numerator /= gcd;
-            Denominator /= gcd;
-        }
-
-        /// <summary>
-        /// Finds greatest common divisor of two numbers
-        /// </summary>
-        /// <param name="a"></param>
-        /// <param name="b"></param>
-        /// <returns></returns>
-        private static int Gcd(int a, int b)
-        {
-            if (a < 0)
-                a = -a;
-            if (b < 0)
-                b = -b;
-
-            while (a != 0 && b != 0)
-            {
-                if (a > b)
-                    a %= b;
-                else
-                    b %= a;
-            }
-
-            return a == 0 ? b : a;
-        }
-
-        /// <summary>
-        /// ToString override to pretty output
-        /// </summary>
-        /// <returns>Pretty string of rational number</returns>
-        public override string ToString()
-        {
-            Simplify();
-            return $"{Numerator} / {Denominator}";
-        }
-
-        public override bool Equals(object obj)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override int GetHashCode()
-        {
-            throw new NotImplementedException();
-        }
+    public override int GetHashCode()
+    {
+        return HashCode.Combine(Numerator, Denominator);
     }
 }
